+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "blk_archive"
+name = "blk_stash"
 version = "0.1.0"
 edition = "2021"
 
@@ -57,5 +57,5 @@ skip_cleanup = []
 debug = 1
 
 [[bin]]
-name = "blk-archive"
+name = "blk-stash"
 path = "src/main.rs"
