+++ doc/Design.md
@@ -1,18 +1,18 @@
 # Introduction
-The thin provisioning device mapper target has always had the facility to take metadata snapshots for use by userland processes.  The intention was this would be used for incremental backups and health monitoring.  I didn't write any tools to use this (other than thin_ls).  The number of third party users is very small; one product uses metadata snaps to do remote replication, and I just discovered [wyng-backup](https://github.com/tasket/wyng-backup), which has similar aims to blk-archive.
+The thin provisioning device mapper target has always had the facility to take metadata snapshots for use by userland processes.  The intention was this would be used for incremental backups and health monitoring.  I didn't write any tools to use this (other than thin_ls).  The number of third party users is very small; one product uses metadata snaps to do remote replication, and I just discovered [wyng-backup](https://github.com/tasket/wyng-backup), which has similar aims to blk-stash.
 
 Meanwhile, some users of thin provisioning are using the thin snapshot facility very frequently to take a form of rolling backup.  This results in thousands of snapshots being present in the pool, resulting in less space and more write amplification as we continually break sharing when performing writes.
 
-blk-archive provides a way to archive any block device, provided it is not in use.  It works particularly well with dm-thin, since thin snapshots can be used to effectively archive live devices.  Metadata snapshots are used to do incremental archiving, and to maximise sharing when restoring a volume to a pool.
+blk-stash provides a way to archive any block device, provided it is not in use.  It works particularly well with dm-thin, since thin snapshots can be used to effectively archive live devices.  Metadata snapshots are used to do incremental archiving, and to maximise sharing when restoring a volume to a pool.
 
 Archived data is deduplicated and compressed then stored within a file system.
 
 The tool is small, probably comprising ~5k lines of code when finished.  Design decisions have been taken to keep things simple.  Giving us less code to maintain, and reducing chances of archive corruption due to software error.  In particular the core archive files are either only appended to or only written once.  By default _pack_ operations are verified to guarantee the stream in the archive matches that of the original device.  The downside of these decisions is we do not support deleting streams from an archive.  This restriction can be ameliorated by migrating streams between archives, which also has the beneficial side effect of defragging streams.
 
-This document describes at the high level the blk-archive implementation.  Read it if you wish to understand how it works.
+This document describes at the high level the blk-stash implementation.  Read it if you wish to understand how it works.
 
 # Slab files
-blk-archive stores its data in _slab files_.  These files consist of a simple header followed by multiple _slabs_.  Each slab is an opaque chunk of data, of a size chosen by the client (typically  4M).   The client may only read an existing slab, or append a new one; there is no way to modify an existing slab.
+blk-stash stores its data in _slab files_.  These files consist of a simple header followed by multiple _slabs_.  Each slab is an opaque chunk of data, of a size chosen by the client (typically  4M).   The client may only read an existing slab, or append a new one; there is no way to modify an existing slab.
 
 Slabs are always checksummed, so corruption at the slab level is detected when the client reads the slab.
 
@@ -27,7 +27,7 @@ There are two classes of data deduplication:
 - Online; deduplicates live data.  This is a _very_ hard problem, having to cope with maintaining large indexes in kernel space, keeping latency of individual IOs low and support random access IO.  Typically fixed size blocks are used.
 - Offline; which generally dedups a sequential stream of data.  This is a much simpler problem.  By maintaining stream locality for the data blocks and hashes it's much more likely that the data we need is already in memory.  The lack of random access IO also means we can use variable block sizes which dramatically improves the amount of duplication found.  Everything is in userland, meaning we don't have to allocate all in core data structures up front.
 
-blk-archive is an offline tool of course.  Two slab files are generated by dedup, one to store the data, and another which stores the hashes + data offsets.
+blk-stash is an offline tool of course.  Two slab files are generated by dedup, one to store the data, and another which stores the hashes + data offsets.
 
 Stream processing goes through these phases:
 
@@ -69,7 +69,7 @@ In addition to the indexer memory the pack process will need a lot of space for 
 When archiving a stream an in core BTree is maintained mapping 256bit hashes to data locations.  This tree doesn't contain all hashes seen, just those for recently visited slabs.  The amount of memory devoted to the hash cache is configurable.  This defaults to 1G.
 
 # Compression
-blk-archive uses zlib with default settings.
+blk-stash uses zlib with default settings.
 
 # Streams
 To reconstruct a stream we need to know which data blocks it is composed of.  Simply writing a list of data slab index + offset to the stream file would consume a lot of space.  So we use various tricks to compress it.
@@ -78,7 +78,7 @@ To reconstruct a stream we need to know which data blocks it is composed of.  Si
 - Unprovisioned regions of a thin volume need their own representation.
 - Surprisingly a lot of blocks contain bytes with identical values.  So we special case these.
 
-Snapshots complicate things.  We know they will share a lot of data entries with the snapshot origin (assuming it's already archived).  One way to take advantage of this is to store references in the stream to an earlier stream, effectively saying something like "the next 63M are identical to stream X and this point".  The downside to this scheme is we start having to read multiple stream files concurrently; potentially very many if people are using a rolling snapshot scenario.  So then we start having to start developing heuristics to say when to use back references, and when to write out the stream in full.  This is too complicated, so instead blk-archive just dedups the stream data itself, and lets the dedup engine find the commonality.
+Snapshots complicate things.  We know they will share a lot of data entries with the snapshot origin (assuming it's already archived).  One way to take advantage of this is to store references in the stream to an earlier stream, effectively saying something like "the next 63M are identical to stream X and this point".  The downside to this scheme is we start having to read multiple stream files concurrently; potentially very many if people are using a rolling snapshot scenario.  So then we start having to start developing heuristics to say when to use back references, and when to write out the stream in full.  This is too complicated, so instead blk-stash just dedups the stream data itself, and lets the dedup engine find the commonality.
 
 The stream is represented as a sequence of instructions for a tiny abstract machine.  The machine has 16 registers arranged in a stack.  Each register consists of a data slab, and offset into the slab.  Some instructions manipulate the stack.  Others adjust the slab or offset fields of the top of the stack.  Finally emit instructions causes runs of data to be emitted and automatically increment the offset field of the top cursor.
