+++ test/vm_test.sh
@@ -31,7 +31,7 @@ PATH=$PATH:$(pwd)/target/release
 export path
 
 if [ ! -d dmtest-python ]; then
-    git clone https://github.com/jthornber/dmtest-python.git || exit 1
+    git clone -b changes_for_rename https://github.com/tasleson/dmtest-python.git || exit 1
 fi
 
 # Create the block devices
@@ -47,7 +47,7 @@ loop3=$(losetup -f --show /block3.img)
 cd "$START_DIR" || exit 1
 cargo test || exit 1
 
-# Run the dmtest-python tests for blk-archive
+# Run the dmtest-python tests for blk-stash
 # Unable to run rolling linux test as we don't have enough disk space in the CI VMs.
 cd "$START_DIR" || exit 1
 # setup the configuration file for dmtest-python
@@ -59,10 +59,10 @@ echo "disable_by_id_check = true" >> config.toml
 
 export DMTEST_RESULT_SET=unit-test
 ./dmtest health || exit 1
-./dmtest run blk-archive/unit/combinations
+./dmtest run blk-slash/unit/combinations
 rc=$?
 if [ $rc -ne 0 ]; then
-    ./dmtest log /blk-archive/unit/combinations
+    ./dmtest log /blk-slash/unit/combinations
     exit 1
 fi
 exit 0
