+++ tests/common/blk_stash.rs
