+++ tests/common/fixture.rs
@@ -3,7 +3,7 @@ use std::os::unix::fs::OpenOptionsExt;
 use std::path::{Path, PathBuf};
 use thinp::file_utils::create_sized_file;
 
-use crate::common::blk_archive::*;
+use crate::common::blk_stash::*;
 use crate::common::block_visitor::*;
 use crate::common::random::Pattern;
 use crate::common::test_dir::*;
